First round to refactor EmojiInteraction out of EmojiPicker
